Revert "Add friends likes"
