+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -1,11 +1,10 @@
 package ru.yandex.practicum.filmorate.controller;
 
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-import ru.yandex.practicum.filmorate.dto.FilmDTO;
 import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.service.film.FilmService;
+import ru.yandex.practicum.filmorate.storage.FilmRepository;
+import ru.yandex.practicum.filmorate.storage.ManagerProvider;
 
 import javax.validation.Valid;
 import java.util.List;
@@ -14,57 +13,26 @@ import java.util.List;
 @RestController
 @RequestMapping("/films")
 public class FilmController {
-    private final FilmService filmService;
-
-    @Autowired
-    public FilmController(FilmService filmService) {
-        this.filmService = filmService;
-    }
+    private final FilmRepository filmsManager = ManagerProvider.getDefaultFilmsManager();
 
     // добавление фильма
     @PostMapping
-    public FilmDTO addFilm(@Valid @RequestBody Film film) {
+    public Film addFilm(@Valid @RequestBody Film film) {
         log.info("Добавление фильма " + film.getName());
-        return filmService.addFilm(film);
+        return filmsManager.addFilm(film);
     }
 
     // обновление фильма
     @PutMapping
-    public FilmDTO updateFilm(@Valid @RequestBody Film updateFilm) {
+    public Film updateFilm(@Valid @RequestBody Film updateFilm) {
         log.info("Обновление фильма " + updateFilm.getName());
-        return filmService.updateFilm(updateFilm);
+        return filmsManager.updateFilm(updateFilm);
     }
 
     // получение списка фильмов
     @GetMapping
-    public List<FilmDTO> getFilms() {
-        log.info("Получение списка всех фильмов " + filmService.getAllFilms());
-        return filmService.getAllFilms();
-    }
-
-    @GetMapping("/{filmId}")
-    public FilmDTO getById(@Valid @PathVariable int filmId) {
-        log.info("Запрос фильма по id=" + filmId);
-        return filmService.getById(filmId);
-    }
-
-    // пользователь ставит лайк фильму
-    @PutMapping("/{filmId}/like/{userId}")
-    public void addLike(@Valid @PathVariable int filmId, @PathVariable int userId) {
-        log.info("Пользователь ставит лайк фильму filmId=" + filmId + " userId=" + userId);
-        filmService.addLike(filmId, userId);
-    }
-
-    // пользователь удаляет лайк
-    @DeleteMapping("/{filmId}/like/{userId}")
-    public void removeLike(@Valid @PathVariable int filmId, @PathVariable int userId) {
-        log.info("Пользователь удаляет лайк фильму filmId=" + filmId + " userId=" + userId);
-        filmService.removeLike(filmId, userId);
-    }
-
-    // возвращает список из первых count фильмов по количеству лайков
-    @GetMapping("/popular")
-    public List<FilmDTO> getPopular(@Valid @RequestParam(required = false, defaultValue = "10") int count) {
-        return filmService.getTop(count);
+    public List<Film> getFilms() {
+        log.info("Получение списка всех фильмов " + filmsManager.getAllFilms());
+        return filmsManager.getAllFilms();
     }
 }
