--- src/main/java/ru/yandex/practicum/filmorate/exception/InternalServerErrorException.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.filmorate.exception;
-
-public class InternalServerErrorException extends RuntimeException {
-    public InternalServerErrorException(String message) {
-        super(message);
-    }
-}
