--- src/main/java/ru/yandex/practicum/filmorate/exception/SearchedObjectNotFoundException.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.filmorate.exception;
-
-public class SearchedObjectNotFoundException extends RuntimeException {
-    public SearchedObjectNotFoundException(String message) {
-        super(message);
-    }
-}
