+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmRepository.java
@@ -1,0 +1,15 @@
+package ru.yandex.practicum.filmorate.storage;
+
+import ru.yandex.practicum.filmorate.model.Film;
+
+import java.util.List;
+
+public interface FilmRepository {
+    Film addFilm(Film film);
+
+    Film updateFilm(Film filmToUpdate);
+
+    List<Film> getAllFilms();
+
+    void clear();
+}
