--- src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -1,23 +1,0 @@
-package ru.yandex.practicum.filmorate.storage.film;
-
-import ru.yandex.practicum.filmorate.model.Film;
-
-import java.util.List;
-
-public interface FilmStorage {
-    Film addFilm(Film film);
-
-    Film updateFilm(Film filmToUpdate);
-
-    List<Film> getAllFilms();
-
-    Film getById(int filmId);
-
-    void addLike(int filmId, int userId);
-
-    void removeLike(int filmId, int userId);
-
-    List<Film> getTop(int count);
-
-    void clear();
-}
